fix:personal datas page changed to loan datas and view in secured loan details loan amount,loan type
